Sprint-8. Работа над замечаниями - 1

+++ EventManagementServiceTests/Infrastructure/DbContextMocker.cs
@@ -35,7 +35,7 @@ public class DbContextMocker()
 
         Mock<ICurrentUserService> currentUserService = new();
         currentUserService.Setup(service => service.UserId).Returns(1);
-        currentUserService.Setup(service => service.Role).Returns((int)UsersRole.User);
+        currentUserService.Setup(service => service.Role).Returns(UsersRole.User);
         currentUserService.Setup(service => service.IsAllowUserOperation(1)).Returns(true);
 
         return new BookingRepository(
@@ -64,7 +64,7 @@ public class DbContextMocker()
 
         Mock<ICurrentUserService> currentUserService = new();
         currentUserService.Setup(service => service.UserId).Returns(1);
-        currentUserService.Setup(service => service.Role).Returns((int)UsersRole.User);
+        currentUserService.Setup(service => service.Role).Returns(UsersRole.User);
         currentUserService.Setup(service => service.IsAllowUserOperation(1)).Returns(true);
 
         IBookingRepository repoBookings = new BookingRepository(
